+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/store/maker/dto/MakerDetailResponse.java
@@ -1,4 +1,4 @@
-package com.drunkenlion.alcoholfriday.domain.admin.store.dto;
+package com.drunkenlion.alcoholfriday.domain.admin.store.maker.dto;
 
 import com.drunkenlion.alcoholfriday.domain.maker.entity.Maker;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -10,13 +10,19 @@ import java.time.LocalDateTime;
 @Builder
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Schema(description = "전체 제조사 조회 항목")
-public class MakerListResponse {
+@Schema(description = "제조사 상세 조회 항목")
+public class MakerDetailResponse {
     @Schema(description = "고유 아이디")
     private Long id;
 
     @Schema(description = "이름")
     private String name;
+
+    @Schema(description = "주소")
+    private String address;
+
+    @Schema(description = "상세주소")
+    private String detail;
 
     @Schema(description = "제조지역")
     private String region;
@@ -24,16 +30,22 @@ public class MakerListResponse {
     @Schema(description = "생성일시")
     private LocalDateTime createdAt;
 
-    @Schema(description = "삭제여부")
-    private boolean deleted;
+    @Schema(description = "마지막 수정일시")
+    private LocalDateTime updatedAt;
 
-    public static MakerListResponse of(Maker maker) {
-        return MakerListResponse.builder()
+    @Schema(description = "삭제일시")
+    private LocalDateTime deletedAt;
+
+    public static MakerDetailResponse of(Maker maker) {
+        return MakerDetailResponse.builder()
                 .id(maker.getId())
                 .name(maker.getName())
+                .address(maker.getAddress())
+                .detail(maker.getDetail())
                 .region(maker.getRegion())
                 .createdAt(maker.getCreatedAt())
-                .deleted(maker.getDeletedAt() != null)
+                .updatedAt(maker.getUpdatedAt())
+                .deletedAt(maker.getDeletedAt())
                 .build();
     }
 }
