+++ src/main/java/com/drunkenlion/alcoholfriday/domain/admin/store/maker/dto/MakerListResponse.java
@@ -1,4 +1,4 @@
-package com.drunkenlion.alcoholfriday.domain.admin.store.dto;
+package com.drunkenlion.alcoholfriday.domain.admin.store.maker.dto;
 
 import com.drunkenlion.alcoholfriday.domain.maker.entity.Maker;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -10,19 +10,13 @@ import java.time.LocalDateTime;
 @Builder
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Schema(description = "제조사 상세 조회 항목")
-public class MakerDetailResponse {
+@Schema(description = "전체 제조사 조회 항목")
+public class MakerListResponse {
     @Schema(description = "고유 아이디")
     private Long id;
 
     @Schema(description = "이름")
     private String name;
-
-    @Schema(description = "주소")
-    private String address;
-
-    @Schema(description = "상세주소")
-    private String detail;
 
     @Schema(description = "제조지역")
     private String region;
@@ -30,22 +24,16 @@ public class MakerDetailResponse {
     @Schema(description = "생성일시")
     private LocalDateTime createdAt;
 
-    @Schema(description = "마지막 수정일시")
-    private LocalDateTime updatedAt;
+    @Schema(description = "삭제여부")
+    private boolean deleted;
 
-    @Schema(description = "삭제일시")
-    private LocalDateTime deletedAt;
-
-    public static MakerDetailResponse of(Maker maker) {
-        return MakerDetailResponse.builder()
+    public static MakerListResponse of(Maker maker) {
+        return MakerListResponse.builder()
                 .id(maker.getId())
                 .name(maker.getName())
-                .address(maker.getAddress())
-                .detail(maker.getDetail())
                 .region(maker.getRegion())
                 .createdAt(maker.getCreatedAt())
-                .updatedAt(maker.getUpdatedAt())
-                .deletedAt(maker.getDeletedAt())
+                .deleted(maker.getDeletedAt() != null)
                 .build();
     }
 }
