+++ src/test/java/com/drunkenlion/alcoholfriday/domain/admin/store/maker/api/AdminStoreMakerControllerTest.java
@@ -1,5 +1,6 @@
-package com.drunkenlion.alcoholfriday.domain.admin.store.api;
+package com.drunkenlion.alcoholfriday.domain.admin.store.maker.api;
 
+import com.drunkenlion.alcoholfriday.domain.admin.store.maker.api.AdminStoreMakerController;
 import com.drunkenlion.alcoholfriday.domain.maker.dao.MakerRepository;
 import com.drunkenlion.alcoholfriday.domain.maker.entity.Maker;
 import com.drunkenlion.alcoholfriday.global.util.TestUtil;
@@ -26,7 +27,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @AutoConfigureMockMvc
 @Transactional
-public class AdminStoreControllerTest {
+public class AdminStoreMakerControllerTest {
     @Autowired
     private MockMvc mvc;
 
@@ -65,7 +66,7 @@ public class AdminStoreControllerTest {
         // then
         resultActions
                 .andExpect(status().isOk())
-                .andExpect(handler().handlerType(AdminStoreController.class))
+                .andExpect(handler().handlerType(AdminStoreMakerController.class))
                 .andExpect(handler().methodName("getMakers"))
                 .andExpect(jsonPath("$.data", instanceOf(List.class)))
                 .andExpect(jsonPath("$.data.length()", is(1)))
@@ -95,7 +96,7 @@ public class AdminStoreControllerTest {
         // then
         resultActions
                 .andExpect(status().isOk())
-                .andExpect(handler().handlerType(AdminStoreController.class))
+                .andExpect(handler().handlerType(AdminStoreMakerController.class))
                 .andExpect(handler().methodName("getMaker"))
                 .andExpect(jsonPath("$", instanceOf(LinkedHashMap.class)))
                 .andExpect(jsonPath("$.id", instanceOf(Number.class)))
@@ -129,7 +130,7 @@ public class AdminStoreControllerTest {
         // then
         resultActions
                 .andExpect(status().isCreated())
-                .andExpect(handler().handlerType(AdminStoreController.class))
+                .andExpect(handler().handlerType(AdminStoreMakerController.class))
                 .andExpect(handler().methodName("createMaker"))
                 .andExpect(jsonPath("$", instanceOf(LinkedHashMap.class)))
                 .andExpect(jsonPath("$.id", instanceOf(Number.class)))
@@ -166,7 +167,7 @@ public class AdminStoreControllerTest {
         // then
         resultActions
                 .andExpect(status().isOk())
-                .andExpect(handler().handlerType(AdminStoreController.class))
+                .andExpect(handler().handlerType(AdminStoreMakerController.class))
                 .andExpect(handler().methodName("modifyMaker"))
                 .andExpect(jsonPath("$", instanceOf(LinkedHashMap.class)))
                 .andExpect(jsonPath("$.id", instanceOf(Number.class)))
@@ -195,7 +196,7 @@ public class AdminStoreControllerTest {
         // then
         resultActions
                 .andExpect(status().isNoContent())
-                .andExpect(handler().handlerType(AdminStoreController.class))
+                .andExpect(handler().handlerType(AdminStoreMakerController.class))
                 .andExpect(handler().methodName("deleteMaker"));
     }
 }
